Refactor Download button and handler to component
