+++ components/form-builder/app/responses/DownloadTable.tsx
@@ -13,8 +13,6 @@ import { DownloadResponseStatus } from "./DownloadResponseStatus";
 import { RemovalStatus } from "./RemovalStatus";
 import { DownloadStatus } from "./DownloadStatus";
 import { useRouter } from "next/router";
-import axios from "axios";
-import { toast } from "../shared/Toast";
 import { useSetting } from "@lib/hooks/useSetting";
 import Link from "next/link";
 import {
@@ -25,16 +23,16 @@ import {
 } from "./DownloadTableReducer";
 import { getDaysPassed } from "@lib/clientHelpers";
 import { Alert } from "@components/globals";
-import { logMessage } from "@lib/logger";
 import {
   CheckAllIcon,
   CheckBoxEmptyIcon,
   CheckIndeterminateIcon,
 } from "@components/form-builder/icons";
+import { DownloadButton } from "./DownloadButton";
 
 interface DownloadTableProps {
   vaultSubmissions: VaultSubmissionList[];
-  formId?: string;
+  formId: string;
   nagwareResult: NagwareResult | null;
   responseDownloadLimit: number;
 }
@@ -139,70 +137,6 @@ export const DownloadTable = ({
     if (nextState.checkedItems.size > 0 && errors.noItemsError) {
       setErrors({ ...errors, noItemsError: false });
     }
-  };
-
-  // NOTE: browsers have different limits for simultaneous downloads. May need to look into
-  // batching file downloads (e.g. 4 at a time) if edge cases/* come up.
-  const handleDownload = async () => {
-    // Reset any errors
-    if (errors.downloadError) {
-      setErrors({ ...errors, downloadError: false });
-    }
-
-    // Handle any errors and show them
-    if (tableItems.checkedItems.size === 0) {
-      if (!errors.noItemsError) {
-        setErrors({ ...errors, noItemsError: true });
-      }
-      return;
-    }
-    if (tableItems.checkedItems.size > MAX_FILE_DOWNLOADS) {
-      if (!errors.maxItemsError) {
-        setErrors({ ...errors, maxItemsError: true });
-      }
-      return;
-    }
-
-    toast.info(
-      t("downloadResponsesTable.notifications.downloadingXFiles", {
-        fileCount: tableItems.checkedItems.size,
-      })
-    );
-
-    const url = `/api/id/${formId}/submissions/download?format=html`;
-    const ids = Array.from(tableItems.checkedItems.keys());
-
-    axios({
-      url,
-      method: "POST",
-      data: {
-        ids: ids.join(","),
-      },
-    })
-      .then((response) => {
-        const interval = 200;
-        response.data.forEach((submission: { id: string; html: string }, i: number) => {
-          setTimeout(() => {
-            const fileName = `${submission.id}.html`;
-            const href = window.URL.createObjectURL(new Blob([submission.html]));
-            const anchorElement = document.createElement("a");
-            anchorElement.href = href;
-            anchorElement.download = fileName;
-            document.body.appendChild(anchorElement);
-            anchorElement.click();
-            document.body.removeChild(anchorElement);
-            window.URL.revokeObjectURL(href);
-          }, interval * i);
-        });
-        setTimeout(() => {
-          router.replace(router.asPath, undefined, { scroll: false });
-          toast.success(t("downloadResponsesTable.notifications.downloadComplete"));
-        }, interval * response.data.length);
-      })
-      .catch((err) => {
-        logMessage.error(err as Error);
-        setErrors({ ...errors, downloadError: true });
-      });
   };
 
   const blockDownload = (
@@ -384,17 +318,14 @@ export const DownloadTable = ({
         </tbody>
       </table>
       <div className="mt-8 flex">
-        <button
-          id="downloadTableButtonId"
-          className="gc-button--blue m-0 w-auto whitespace-nowrap"
-          type="button"
-          onClick={handleDownload}
-          aria-live="polite"
-        >
-          {t("downloadResponsesTable.downloadXSelectedResponses", {
-            size: tableItems.checkedItems.size,
-          })}
-        </button>
+        <DownloadButton
+          formId={formId}
+          router={router}
+          setErrors={setErrors}
+          errors={errors}
+          tableItems={tableItems}
+          responseDownloadLimit={MAX_FILE_DOWNLOADS}
+        />
 
         <div id="notificationsBottom" className="ml-4">
           {tableItems.checkedItems.size > MAX_FILE_DOWNLOADS && (
